move code to handle strand bias to `the strand_bias_handler.py` module
